TOS setup: add README on setting up TOS and hooks to populate data.
The workbench-tos application is required by the firecloud UI to allow users to register for the app. Unlike the rest of firecloud, TOS is deployed as a cloud function.

This commit adds instructions and populate / unpopulate commands to allow the TOS cloud function to be initialized in the FIAB GCP project.